- Fixes - Move render_bloglist_embed function to blogs.embeds - Use deepcopy for bloglist assets dict. - Add blogs list embed publishing settings to main settings.py
